+++ js/form.js
@@ -1,0 +1,4 @@
+import {verifyStrLength} from '/js/util.js';
+
+//Заглушка для линтера
+verifyStrLength('Какой-то текст...', 50);
